requirements.txt
gradio
requests
unidecode
python-dotenv
